requirements.txt
DjangDjango >= 4.2 , <5.0
python-decouple == 3.8
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psycop2-binary==2.9.9